Fix type annotation for LineItemType enum
The annotation says it would expect an instance of LineItemType. But it expects a string (a constant from the LineItemType).
This currently fails any autocomplete and static analysis.

Using LineItemType::* fixes this as allowed by phpstan, see https://phpstan.org/writing-php-code/phpdoc-types#literals-and-constants

But in general the same problem seems to apply to all enums in this library, e.g. also TransactionState.
Is this generated code? Then it needs to be fixed in the generator I assume.